First pass refactor of ingest into multiple threads with telescope model
Have refactored ingest into a CBF and TM ingest thread system. These
both use a TelescopeModel which is initialised based on the current
subarray in use.

The model gets populated and kept in memory during the experiment,
on capture done it is written out to the HDF5 file.

Many bugs remain, but this does build, run and accept sensor data.