Made TelescopeModel the authority for the HDF5 structure and more !
The biggest change is pushing almost all file writing functionality
into the model. The model now has the version number and is basically
the authority for file structure. correlator_data and timestamps
are the two obvious exceptions to this rule, and perhaps require
some thought.

Also started on the path to better simulation capabalities by providing
a mechanism for a Model to replay itself by generation appropriate
SPEAD heaps. The plan is to add an HDF5->Model method as well and
then we can do HDF5->Model->SPEAD->Model->HDF5, which would be
cool :)