TelescopeModel can now be initialised from an existing HDF5 file.
Allows an existing HDF5 file to be used to populate a model with
components, attributes and sensors.

Relies on a 'class' attribute of the component group to figure
out which class to use to build the component. This class needs
to be in the global namespace, which generally means that you should
keep the component class definitions in the same file as the
TelescopeModel class :)