Fix support for obs_label and obs_params sensors
The observation labels and parameters now appear as sensors on a
virtual 'obs' CAM device and their full names will therefore be
'obs_label' and 'obs_params'. The TelescopeModel is updated to
require these sensors on the Observation component. The KATCP
requests on cam2spead are updated to set_obs_label and
set_obs_param (singular because you set one parameter at a time).

Also fix sensor string update for obs_params that used to contain
two quoted strings to a single quoted string with a space separating
the parameter key and value. This is because the value string will
be eval()ed in ingest and having two quoted strings breaks it.