First attempt at a FakeClient object
This a rough first draft (after several refactors, of course), using
a "face hugger" approach. The AntennaModel is a standalone object
that is blissfully unaware of the client request/sensor interface.
Variables with names that match expected sensors will automatically
pass value changes to the corresponding sensors. Methods with names
starting with 'req_' will be exposed as requests on the client.