Add support for model attributes
The config file now contains both sensors and attributes, as well as
a top-level [Telescope] section that names the various components
in the telescope system. Sensors and attributes may depend on the
component instance name. Attributes naturally match the arguments
passed to Model.__init__(), so that a FakeClient can be constructed
based only on the model config file.