No need for an explicit stop anymore
When used in a script, a user would use the context manager interface
of FakeTelescope (i.e. in a with statement), which stops and joins
the updater thread. When used in an interactive shell, a user does
not have to stop anything, as the updater thread is now daemonic.

Neilen's concern is noted: it is useful to stop() and join() separately
to speed up situations like tests which could start up multiple
FakeTelescopes. If done in one step, the join() overhead is multiplied
by the number of threads, while separating them allows threads to stop
in parallel in the background. This is now not an issue for the user
anymore, as there is no need for explicit stopping, while a test
writer can explicitly call self.updater.stop() and self.updater.join()
to save time.