Don't set the target unnecessarily
Don't set the target unnecessarily as this upsets the antenna proxy,
causing momentary loss of lock as a STOP is issued even if the target
remains the same. Check each antenna individually as they may have
different targets when they join a new sub-array. This addresses JIRA
issue MRTS-4 and Mantis #3201.

git-svn-id: https://katfs.kat.ac.za/svnDS/code/katcorelib/trunk@20133 b56c7a7d-ac47-0410-a905-b6f7de78896a
git-svn-id: https://katfs.kat.ac.za/svnDS/code/katcorelib/trunk@20856 b56c7a7d-ac47-0410-a905-b6f7de78896a
git-svn-id: https://katfs.kat.ac.za/svnDS/code/katcorelib/trunk@21111 b56c7a7d-ac47-0410-a905-b6f7de78896a