Move script log handling from ingest to cam2spead
Update the script log via a (virtual) observation sensor on cam2spead
instead of via a special call to the CBF ingest thread. The rationale
is that the script log is something that produces events while the
observation script is running similar to other CAM sensors. This ensures
that third-party users will get their script_log fix via SPEAD as well.
The script_log becomes an optional sensor of the obs telescope component.