Improve DBE mode change
Synchronise kat7_session better with dbe7 proxy during mode change by waiting
on mode-status sensor instead of dbe-mode sensor. The latter is problematic
when the mode is set to the same value it had before, which means that the
wait() returns immediately while the dbe7 proxy is still busy with the mode()
command.

git-svn-id: https://katfs.kat.ac.za/svnDS/code/katcorelib/branches/karoocamv8/@21382 b56c7a7d-ac47-0410-a905-b6f7de78896a
git-svn-id: https://katfs.kat.ac.za/svnDS/code/katcorelib/branches/karoocamv8/@21383 b56c7a7d-ac47-0410-a905-b6f7de78896a
git-svn-id: https://katfs.kat.ac.za/svnDS/code/katcorelib/branches/karoocamv8/@21384 b56c7a7d-ac47-0410-a905-b6f7de78896a
git-svn-id: https://katfs.kat.ac.za/svnDS/code/katcorelib/trunk@21385 b56c7a7d-ac47-0410-a905-b6f7de78896a