Fix ObsParams to make .update(dict) work
The ObsParams class successfully overrode the dict.__setitem__ method but
ObsParams.update is still the original dict.update which ignores the custom
__setitem__. This meant that additional script options like --mode did not
make it into the SPEAD stream. The moral: don't derive from dict... :-)

Rewrite the class to be based on the collections.MutableMapping abstract
base class instead. This ensures that dict is properly extended by containing
a dict inside ObsParams instead of deriving from it. Let __setitem__ and
__delitem__ communicate with the SPEAD stream, and the public methods are
provided for free.

git-svn-id: https://katfs.kat.ac.za/svnDS/code/katcorelib/trunk@21784 b56c7a7d-ac47-0410-a905-b6f7de78896a