Improve docstrings and do minor API tweaks
Get docstrings in standard NumPy shape. Rename KATResource.sources to
KATResource.targets to make it more generic as well as better aligned with the
terminology in katpoint. Rename KATResourceClient to KATClientResource to
reflect the fact that it is a resource in the first place. Move StatusTuple
documentation to the named tuple itself, using the class trick suggested in
http://stackoverflow.com/questions/1606436.