Generalise a fake component
Given the pending distinction between fake, sdp (simulator) and cam (real)
components, the fake component base class is moved out to the top and
renamed to TelstateUpdatingComponent, which is exactly its purpose. This
allows non-fake components to use this functionality too.

The Observation component will be found unmodified in the other systems,
for example. It is moved to a yet-to-be-committed file.