New MultiComponent
This component combines a collection of similar components, passing attribute
assignments and method calls through to each element. Getting attributes back
is not supported yet as this is tricky if the underlying attributes differ per
element. One way out would be to pick a reference element (the first one?).

This is used to implement receptor groups when constructing from a config file.