New TargetObserverMixin
This mixin adds target and observer properties to any component. The two
properties are related in that the observer is the default katpoint.Antenna
object for the katpoint.Target object. Redo the fake and sdp components that
need these.

Also move sdp.Observation back to fake and import it into sdp instead, as this
component is more typical of a fake component.

Pick default attribute values for the fake.CorrelatorBeamformer based on
the default product of c856M4k and dump rate of 2 Hz.

Only update AntennaPositioner target if it differs.