Simplify default session configuration
It is now possible to make a dummy CaptureSession without any components,
that can be used to parse basic command line arguments. This is used to
extract the --config argument in session_from_commandline, which leaves
a single place to update the default config location in session.py
(previously there were two - that is, one too many).

Make a small fake RTS the default config, as this avoids connecting to
actual lab services while others are using it.