Add --dont-stop arg to preserve session at end
When run with --dont-stop, leave the session running after the script is done.
In other words, don't product_deconfigure and don't stop components or the
ioloop. This leaves the SDP pipeline and telstate accessible for debugging
purposes. Refactor starting and stopping to facilitate this. Just call _stop()
to close the session when done.

Also initialise obs_params with the original script arguments.