Explicitly set targets parameter in obs script
Since CaptureSession.targets now defaults to False, it has to be set in
observation scripts that operate on targets. Instead of setting it on the
session object itself in the script (which seems a little clunky), pass it
parameter to the workhorse session_from_commandline(). There is no need to
propagate it to the parameter lists of CaptureSession and session_from_config
as well. The basic_track test now takes an actual target and observes it.