Plumb in updater thread to get dry-run working
This introduces the --dry-run option which, together with WarpClock and
PeriodicUpdaterThread, runs the script faster than realtime. Patch the track()
command to use this, so that basic_track works with --dry-run. This speeds up
unit tests too. Dry running is only enabled if all components are fake and
support updating.