Rework logging configuration for CaptureSession
All logging-related functionality shifts to its own module. The logging module
is monkey-patched to enable the use of a custom clock for all log timestamps.
This makes *all* logs appear at the predicted times during dry running (not
just the ones emanating from the CaptureSession logger). The ScriptLogHandler turns into a more generic RobustDeliveryHandler. The logging system is configured via a LoggingConfigurer object that adds the appropriate handlers
to the root logger, sets its level and formatting and also sets up the custom
clock. These steps are undone in the restore() step of the configurer, which will be useful once we embed dry-run CaptureSessions within normal sessions.