Simplify logging configuration by removing state
Instead of storing the old state in an object in order to restore the logging
system, rather just reconfigure the system again via a stateless function call.
The only parameter left is the log level, which initially comes from args and
can later be obtained from obs_params. This now also allows the in-place
modification of an existing script log handler, allowing sessions to be nested
without duplicating logs (just call configure_logging again in the outer
session when the inner session is done).