Make CaptureState an enum34.Enum
This is finally an option as katmisc now also uses enum34 (the old
enum package would have clashed when combining session and kattelmod,
me tries to remember). Go for a straightforward IntEnum as opposed to
rolling an OrderedEnum as per the enum34 documentation, since there
is little chance of comparing CaptureState with another IntEnum.