Flesh out collect_targets
This is an adaptation of the original `collect_targets` function in
`observe.py`. It handles catalogues and description strings but not
yet names, since that requires a default catalogue somewhere. This
is so that `session.targets` can be a Catalogue that can be filtered
and such in the observation script.