Fix "coroutine was never awaited" warnings
These occurred if CaptureSession.connect failed, and hence the run
function was never reached. Instead of taking a coroutine, `run` now
takes a coroutine function (aka `async def`), so that the coroutine
object is only created when it is ready to be awaited. This has the side
benefit of now having to pass session and args twice in the run call.