Fix unit tests on Python 3.7+
When the periodic updater falls behind, it was passing a negative value
to asyncio.sleep. From Python 3.7+ that takes the same fast path as
sleeping for 0 seconds, which has the effect of waking back up again
before timer callbacks are triggered.

To make behaviour consistent, force negative values to zero to get the
Python 3.7+ behaviour all the time, and update the test to match.

Closes SPR1-651.