SPR1-2828 Base WarpEventLoopTestCase on fixtures
This reduces dependence on asynctest by introducing loop and policy
fixtures instead.

The `event_loop` fixture overrides the standard one provided by
`pytest-asyncio` within the `WarpEventLoopTestCase` class. It is
effectively an autouse fixture and therefore akin to setup_method.
Setting the policy in `setup_class` as before did not work because
`event_loop` ran before that and ended up using a standard non-warp
loop.

Use an autouse fixture to create an async setup_method (not yet
supported by Pytest). Reduce state by asking for the running loop
when we need it, instead of storing it on the test object.

Process log messages via the caplog fixture.