SPR1-2828 Replace asynctest with async-solipsism
Replace the last piece of asynctest, which is the `ClockedTestCase`.
@bmerry's async-solipsism is even nicer and already KATCP friendly
with its support of `asyncio.start_server` and other interactions.
While a ClockedTestCase needs to be advance()d, a solipsist event loop
just needs some sleep.

The async setup function becomes an autouse yield fixture, mostly to
handle cleanup. The cleanup (especially of the writer) took a while
to figure out and might still be improved. Get the running loop from
asyncio instead of storing it on the object.

The synchronous `test_bad_katcp_endpoint` did not play nicely with
solipsism (SleepForeverError), so move it out of the class.

This took some inspiration from aiokatcp's test_client.py.