SPR1-2770 Introduce ComponentNotReadyError
The existing custom exceptions were a bit confusing.

ConnectionError implies a network error and is also a standard library
error these days. However, in this context it is more likely to occur
when components are started and stopped in the wrong order.

Rather make these two a single custom error for now until we see
a need to separate them again.