SPR1-2376 Fix major bug in fake system
Instead of returning the initial capture state, the fake product
configure returns the final capture state. This is disastrous as it
bypasses all capture phases in CaptureSession (since it thinks it is
already capturing). This hid a lot of other bugs, since very little of
the machinery was exercised. It's a miracle that this package did
anything useful in 7 years :`-D.