SPR1-2376 Fix race condition on telstate flushes
`TelstateUpdatingComponent._flush()` is called in two main places:
by the component `_start` and by the periodic updater. Consider an
update queue with a single task. The component and the updater can
both determine that the queue is not empty and end up awaiting on the
same task. The first one done will pop the task afterwards, causing
a crash when the second one attempts to pop an empty queue.

Rather pop the task up front and then await it, so that checking and
modifying the queue happen atomically.