SPR1-2376 Use memory-backed telstate for fakes
Rewrite the session telstate machinery. Promote the attribute to
`CaptureSession.telstate` and ensure that it will always be a valid
`TelescopeState` object (no more None). Let the fake SDP return a
'fake' endpoint (not sure how I missed that one...). The only other
valid value for an endpoint is a proper host:port spec (no more '').
The real SDP therefore raises `ComponentNotReadyError` if the endpoint
is unavailable.

The session (and all components) now start out with a "fake" telstate
with memory backend. This may be replaced with a Redis telstate during
product configuration, but it is already sufficient for fake systems.
Upgrade the entire session telstate to the capture block view during
capture init (not just the obs telstate).

I considered initializing each TelstateUpdatingComponent with
a memory-backed telstate to avoid it being None but this is less useful.
It will create several unrelated telstates, while setting it to None
still allows the telstate updates to be disabled (useful for e.g. obs).