SPR1-2376 Add dedicated CaptureSession test
This runs on a WarpEventLoop. I considered using a solipsist loop,
but that would require rebasing _WarpClock on the solipsist Clock
(a nice idea but the solipsist loop only really helps for tests
as it is very shy).

Use fixtures extensively to override the event loop and create the
session and associated argparser. Extend `session_from_commandline`
to take arguments directly instead of via sys.argv, which is useful
for tests. This was the simplest way to make a CaptureSession with
targets and custom arguments.

Check that telstate actually contains some of the expected obs keys.
This uncovered a lot of bugs, now fixed in the preceding commits.
Check that the target is correct.

Code coverage increased from 55% to 82%. :-)